requirements.txt
sentence-transformers
faiss-cpu
python-dotenv
groq